+++ app/Http/Controllers/AvatarController.php
@@ -123,7 +123,7 @@ class AvatarController extends Controller
         if( $avatar->media_path == 'public/avatars/default.png' || 
             $avatar->thumb_path == 'public/avatars/default.png' ||
             $avatar->media_path == 'public/avatars/default.jpg' || 
-            $avatar->thumb_path == 'public/avatars/default.jpg' ||
+            $avatar->thumb_path == 'public/avatars/default.jpg'
 
         ) {
             return;