+++ app/Jobs/StatusPipeline/StatusEntityLexer.php
@@ -52,6 +52,17 @@ class StatusEntityLexer implements ShouldQueue
     public function handle()
     {
         $profile = $this->status->profile;
+
+        $count = $profile->statuses()
+        ->getQuery()
+        ->whereIn('type', ['photo', 'photo:album', 'video', 'video:album', 'photo:video:album'])
+        ->whereNull('in_reply_to_id')
+        ->whereNull('reblog_of_id')
+        ->count();
+
+        $profile->status_count = $count;
+        $profile->save();
+
         if($profile->no_autolink == false) {
             $this->parseEntities();
         }